% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-building.R
\name{build_imputation_units_single}
\alias{build_imputation_units_single}
\title{Build imputation units}
\usage{
build_imputation_units_single(ev)
}
\arguments{
\item{ev}{The \code{ev} member of a \code{co_events_frame} object}
}
\description{
An imputation unit is a set of one or more pairwise overlapping
intervals on which the event history may be imputed
independently of all intervals outside of their union.
}
