% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{co_events}
\alias{co_events}
\title{Create \code{co_events} object.}
\usage{
co_events(
  data_covariates,
  data_events,
  id,
  t_start,
  t_end,
  e_min,
  e_max,
  fill = NA
)
}
\arguments{
\item{data_covariates}{A data frame containing time-varying covariates
over disjoint intervals within units,
one row per interval, including columns for
unit id, interval start, and interval end.}

\item{data_events}{A data frame containing censored event counts
in intervals possibly overlapping within units,
one row per interval, including columns for
unit id and minimum and maximum event counts.}

\item{id}{The column in \code{data_covariates} and \code{data_events}
containing IDs linking units between the datasets.}

\item{t_start}{The column in \code{data_covariates} and \code{data_events}
containing the (open) start time of each interval.}

\item{t_end}{The column in \code{data_covariates} and \code{data_events}
containing the (closed) end time of each interval.}

\item{e_min}{The column in \code{data_events}
containing the minimum number of events in each interval.}

\item{e_max}{The column in \code{data_events}
containing the maximum number of events in each interval.}

\item{fill}{A named list specifying the value of each time-varying covariate
that should be used for newly added rows.}
}
\value{
An object of class \code{co_events}.
A \code{co_events} object is a list indexed by IDs from the source
data objects. Each list element contains \code{covariates} and
\code{events} elements, which are data frames comprised of the rows
of \code{data_covariates} and \code{data_events}, respectively,
corresponding to the ID named by the list index.
The \code{co_events} object also has the attribute
\code{"special_cols"}, a named vector with components:
\describe{
\item{\code{id}}{character string indicating name of the column
passed as the \code{id} argument;}
\item{\code{t_start}}{character string indicating name of the column
passed as the \code{t_start} argument;}
\item{\code{t_end}}{character string indicating name of the column
passed as the \code{t_end} argument;}
\item{\code{e_min}}{character string indicating name of the column
passed as the \code{e_min} argument;}
\item{\code{e_max}}{character string indicating name of the column
passed as the \code{e_max} argument.}
}
}
\description{
Combine time-varying covariate data and censored event data
into a single computation-friendly object.
}
