% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-building.R
\name{co_events_frame}
\alias{co_events_frame}
\title{Build model frames from \code{co_events} object.}
\usage{
co_events_frame(co_events, formula, contrasts = NULL, ...)
}
\arguments{
\item{co_events}{A \code{co_events} object.}

\item{formula}{An object of class \code{formula} symbolically describing
the model to be fit.}

\item{contrasts}{An optional list to be passed to \code{model.matrix}.}

\item{...}{Additional arguments to be passed to \code{model.frame}.}
}
\value{
An object of class \code{co_events_frame}.
In addition to the components of the given \code{co_events} object,
each list element has the following sub-elements:
\describe{
\item{\code{X}}{a model matrix built from \code{covariates}
and the \code{formula} argument;}
\item{\code{cov_t}}{the \code{t_start} and \code{t_end}
columns of \code{covariates} but with standardized variable
names \code{"t_start"} and \code{"t_end"};}
\item{\code{ev}}{same as the \code{events} component
but with standardized variable names \code{"t_start"},
\code{"t_end"}, \code{"e_min"}, and \code{"e_max"}.}
}
}
\description{
For each element in a \code{co_events} object,
builds a model frame from \code{covariates}, and constructs
data frames for intervals with standardized column names
to be used internally in model fitting functions.
}
