% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{collapse_interval_data}
\alias{collapse_interval_data}
\title{Collapse redundant intervals in dataset}
\usage{
collapse_interval_data(data, t_start_var, t_end_var)
}
\arguments{
\item{data}{A data frame.}

\item{t_start_var}{A string indicating the name of the column in \code{data}
storing the start of each interval.}

\item{t_end_var}{A string indicating the name of the column in \code{data}
storing the start of each interval.}
}
\value{
A data frame in which adjacent or overlapping intervals with
identical data combined into one interval.
}
\description{
Combines adjacent or overlapping intervals with identical
data into one interval.
}
\examples{
data_overlap_ints <- data.frame(
id = c(1, 1, 1, 1,
       2, 2, 2, 2),
t_start = c(0, 4, 7, 21,
            7, 14, 21, 28),
t_end = c(7, 11, 14, 28,
          14, 21, 28, 35),
med = c("A", "A", "B", "B",
        "A", "A", "A", "A")
)

data_overlap_ints \%>\% collapse_interval_data("t_start", "t_end")

}
