% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{complete_interval_data}
\alias{complete_interval_data}
\title{Fill in gaps between intervals}
\usage{
complete_interval_data(
  data,
  id,
  t_start_var,
  t_end_var,
  fill = NA,
  new_nodes = c()
)
}
\arguments{
\item{data}{A data frame.}

\item{t_start_var}{The column in \code{data}
storing the start of each interval.}

\item{t_end_var}{The column in \code{data}
storing the end of each interval.}

\item{fill}{A named list specifying the value of each remaining variable
that should be used for newly added rows.}

\item{new_nodes}{A vector of new interval endpoints that should
be added regardless of whether they appear in the data.}
}
\value{
A data frame containing the original intervals plus
the intervals implied by the gaps between them.
}
\description{
Creates additional data rows representing
gaps between existing intervals.
}
