% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{complete_interval_data_single}
\alias{complete_interval_data_single}
\title{Single-id helper for \code{complete_interval_data}}
\usage{
complete_interval_data_single(data, t_start, t_end, fill = NA, new_nodes = c())
}
\description{
Single-id helper for \code{complete_interval_data}
}
