% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{expect_cum_weibull}
\alias{expect_cum_weibull}
\title{Weibull expected cumulative event count}
\usage{
expect_cum_weibull(t, shape = 1, scale = 1, k = shape, b = scale^(-shape), ...)
}
\arguments{
\item{t}{Time at which to evaluate expected cumulative count.}

\item{shape}{Shape of Weibull distribution, as in \code{stats::Weibull}.}

\item{scale}{Scale of Weibull distribution, as in \code{stats::Weibull}.}

\item{k}{Alternative parameterization (shape).}

\item{b}{Alternative parameterization (scale).}
}
\value{
Expected cumulative event count at time \code{t}.
}
\description{
Expected cumulative event count of Poisson process
with Weibull hazard.
Default parameterization matches that of \code{stats::Weibull}.
}
\examples{
expect_cum_weibull(4, 2, 3)
expect_cum_weibull(4, k = 2, b = 1 / 3^2)
}
