% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{expect_cum_weibull_inverse}
\alias{expect_cum_weibull_inverse}
\title{Inverse of Weibull expected cumulative event count}
\usage{
expect_cum_weibull_inverse(
  m,
  shape = 1,
  scale = 1,
  k = shape,
  b = scale^(-shape),
  ...
)
}
\arguments{
\item{m}{Expected (mean) Weibull cumulative event count.}

\item{shape}{Shape of Weibull distribution, as in \code{stats::Weibull}.}

\item{scale}{Scale of Weibull distribution, as in \code{stats::Weibull}.}

\item{k}{Alternative parameterization (shape).}

\item{b}{Alternative parameterization (scale).}
}
\value{
Time at which Weibull expected cumulative event count equals \code{m}.
}
\description{
Inverse of expected cumulative event count of Poisson process
with Weibull hazard.
Default parameterization matches that of \code{stats::Weibull}.
}
\examples{
expect_cum_weibull_inverse(expect_cum_weibull(4, 2, 3))
}
