% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{expect_cum_weibull_tvc_inverse}
\alias{expect_cum_weibull_tvc_inverse}
\title{Inverse of Weibull expected cumulative event count with time-varying covariates}
\usage{
expect_cum_weibull_tvc_inverse(
  m,
  lin_pred = 0,
  t_breaks = Inf,
  shape = 1,
  scale = 1,
  k = shape,
  b = scale^(-shape),
  ...
)
}
\arguments{
\item{m}{Expected (mean) Weibull cumulative event count.}

\item{lin_pred}{Vector of log scaling factors on baseline expected
cumulative event count function over intervals.}

\item{t_breaks}{Vector of right (closed) endpoints of intervals
over which lin_pred applies.
Elements must be strictly increasing and at least one
must be \code{>= t}.}

\item{shape}{Shape of Weibull distribution, as in \code{stats::Weibull}.}

\item{scale}{Scale of Weibull distribution, as in \code{stats::Weibull}.}

\item{k}{Alternative parameterization (shape).}

\item{b}{Alternative parameterization (scale).}
}
\value{
Time at which Weibull expected cumulative event count equals \code{m}.
}
\description{
Inverse of expected cumulative event count
given Weibull baseline expected cumulative event count and
piecewise constant multiplicative factor exp(lin_pred).
}
\examples{
expect_cum_weibull_tvc_inverse(
  expect_cum_weibull_tvc(3, lin_pred = 1 : 5, t_breaks = 1 : 5, k = 1, b = 1),
  lin_pred = 1 : 5, t_breaks = 1 : 5, k = 1, b = 1)

}
