% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{merge_interval_data}
\alias{merge_interval_data}
\title{Merge new time-dependent variable into interval dataset}
\usage{
merge_interval_data(
  data,
  new_data,
  id,
  t_start_var,
  t_end_var,
  new_var,
  fill = NA
)
}
\arguments{
\item{data}{A data frame.}

\item{new_data}{A data frame containing the new variable to be merged.}

\item{t_start_var}{The column in \code{data} and \code{new_data}
storing the start of each interval.}

\item{t_end_var}{The column in \code{data} and \code{new_data}
storing the end of each interval.}

\item{new_var}{The column in \code{new_data} to be merged into \code{data}.}

\item{fill}{A named list specifying the value of each remaining variable
that should be used for newly added rows.}
}
\value{
A data frame containing the original columns of \code{data} plus
the new column from \code{new_data} with subdivided intervals
to accommodate all variables.
}
\description{
Merges an additional column from a new dataset
into an existing interval dataset,
subdividing intervals as necessary.
}
