% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{merge_interval_data_single}
\alias{merge_interval_data_single}
\title{Single-id helper for \code{merge_interval_data}}
\usage{
merge_interval_data_single(data, new_data, t_start, t_end, new_var, fill = NA)
}
\description{
Single-id helper for \code{merge_interval_data}
}
