% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{rpois_trunc}
\alias{rpois_trunc}
\title{Truncated Poisson generation}
\usage{
rpois_trunc(
  n,
  lambda,
  min = 0,
  max = Inf,
  parallel_draws = n * 2,
  max_tries = 1000
)
}
\arguments{
\item{n}{Number of random values to return.}

\item{lambda}{Non-negative (scalar) mean of un-truncated distribution.}

\item{min}{The minimum allowed return value.}

\item{max}{The maximum allowed return value.}

\item{parallel_draws}{For left-only-truncated draws, the number of
simultaneous draws by the rejection sampler.}

\item{max_tries}{For left-only-truncated draws, the maximum number of
iterations of the rejection sampler.}
}
\value{
A vector of draws.
}
\description{
Generates random draws from un-, left-, right-, or double-truncated
Poisson distribution with parameter \code{lambda} truncated to \code{[min, max]}
(inclusive). The Poisson distribution is parameterized identically to
\code{stats::rpois}, but \code{lambda} cannot be a vector.
}
\examples{
rpois_trunc(100, lambda = 1)
rpois_trunc(100, lambda = 1, min = 1)
rpois_trunc(100, lambda = 1, max = 10)
rpois_trunc(100, lambda = 1, min = 1, max = 10)
}
